requirements.txt
Flask
markdown
pyyaml
